Add files via upload
Do you want to hide and disable the comment author link in WordPress and hide the real user author name from the blog post? This is a plugin that is minimalist with not much code; only PHP script and little CSS will be inlined to ensure that the author is no longer visible to the online public.